Set current analytical object
