+++ src/util/analytics.js
@@ -3,7 +3,6 @@ import { sortBy, negate } from 'lodash/fp';
 import { isValidUid } from 'd2/uid';
 import { periodNames } from '../constants/periods';
 import { dimConf } from '../constants/dimension';
-import { loadDataItemLegendSet } from '../util/legend';
 
 /* DIMENSIONS */
 
@@ -246,80 +245,3 @@ export const getApiResponseNames = ({ metaData, headers }) => ({
     true: i18n.t('Yes'),
     false: i18n.t('No'),
 });
-
-// Returns the data items of the first dx dimension in an analytical object
-export const getDataDimensionsFromAnalyticalObject = ao => {
-    const { columns, rows, filters } = ao;
-
-    // TODO: Should filters be included?
-    const dataDim = [...columns, ...rows, ...filters].find(
-        i => i.dimension === 'dx'
-    );
-
-    // We only use the first dx dimension
-    return dataDim ? dataDim.items : [];
-};
-
-// Checks if anaytical object is valid as a map layer
-export const isValidAnalyticalObject = ao => {
-    const { columns, rows, filters } = ao;
-    const dimensions = [...columns, ...rows, ...filters];
-    const dataItems = dimensions.filter(i => i.dimension === 'dx');
-
-    return dataItems.length === 1 && dataItems[0].items.length === 1;
-};
-
-// Returns a thematic layer config from an analytical object
-export const getThematicLayerFromAnalyticalObject = async (
-    ao,
-    dataId,
-    isVisible = true
-) => {
-    const { columns, rows, filters, yearlySeries, aggregationType } = ao;
-    const dimensions = [...columns, ...rows, ...filters];
-    const dataDims = getDataDimensionsFromAnalyticalObject(ao);
-    const orgUnits = dimensions.find(i => i.dimension === 'ou');
-    let period = dimensions.find(i => i.dimension === 'pe');
-    let dataDim = dataDims[0];
-
-    if (dataId) {
-        dataDim = dataDims.find(item => item.id === dataId);
-    }
-
-    // Load default legend set for selected data diemension
-    const legendSet = await loadDataItemLegendSet(dataDim);
-
-    // Currently we only support one period in map filters so we select the first
-    if (yearlySeries && yearlySeries.length) {
-        period = {
-            dimension: 'pe',
-            items: [
-                {
-                    id: yearlySeries[0],
-                    name: getPeriodNameFromId(yearlySeries[0]),
-                },
-            ],
-        };
-    }
-
-    return {
-        layer: 'thematic',
-        columns: [{ dimension: 'dx', items: [dataDim] }],
-        rows: [orgUnits],
-        filters: [period],
-        aggregationType,
-        legendSet,
-        isVisible,
-    };
-};
-
-export const getAnalyticalObjectFromThematicLayer = layer => {
-    const { columns, rows, filters, aggregationType } = layer;
-
-    return {
-        columns,
-        rows,
-        filters,
-        aggregationType,
-    };
-};
